[apnsd] More clever logic to check if error responses.
Before, APNS lagged a little and failed to return the error response
in "max_error_wait" seconds, the APNSAgent thread was block on the
notification queue.  If no new notification comes in, we may wait a
long time before noticing the error.

This commit implements an increasing timeout while blocking on the
notification queue.  After a while, we simply give up and block
forever on the notification queue.