[apnsd] Implement FeedbackAgent.
Unfortunately, this development has been tested because I don't have
any device to test it.

The "test" socket opened to the feedback service after reading the
configuration file is passed to the FeedbackAgent thread because the
feedback service may write feedback tuples immediately on it and we
don't want to lose it.