[push2mob] Replace PersisitentQueue with PersistentFIFO.
PersistentQueue was derived from the Queue.Queue class while PersistentFIFO
derives from OrderedPersistentQueue.

This doesn't bring anything functionally for now but reduces code size.
In the later commit, we will be able to use the rowid returned by the
put() function to generate a notification ID for APNS, instead of
maintaining our own counter in a separate table.