[push2mob] In OrderedPersistentQueue, lock the condvar in batch mode.
This is an attempt to workaround a segfault in pysqlite3.so and
other myterious behaviours.

In don't really understand the problem but pysqlite3 tries to keep track
of transaction state and seems to lose itself when multiple threads use
the same connection to perform various tasks within an explicit transaction.

This commit just acquire the lock when a batch insert is started and
release it when the batch inserts is ended.  The put() operation has been
modified to not acquire the lock in the middle of a batch.  The other
operations still acquire it, so they just wait.