[push2mob] Create one database connection per thread.
pysqlite3 tries to keep track of transactions.  It seems there is a bug
when an explicit transaction is opened and a single connection is
shared by multiple threads.

The patch creates a single connection for each thread.  Besides, it
prevents concurrent access to each table with an application-level
lock.  This probably reduces concurrency but at least it works...

Note for later when I will have more free time: report the bug to the
pysqlite3 project.