[push2mob] Improve timely queue.
Before we hacked around the derived Queue object to achieve this but this
was not perfect.

This commit implement a similar but stripped down interface to the Queue
object (thus "SQueue") that does the right thing.

Given the latter doesn't derive from CheckpointableQueue any more, extract
the Checkpointable interface into its own class.  CheckpointableQueue is
just a class inheriting from Checkpointable and Queue now :).

Remove the Reminder which was part of the hack.